+++ client/src/app/Appraisals/title.jsx
@@ -1,17 +1,18 @@
 
 
 export default function Title(){
+  return(
     <div className="mx-auto max-w-2xl lg:text-center">
-          <h2 className="text-base font-semibold leading-7 text-indigo-600">
-            Nosotros
-          </h2>
-          <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-            Formulario Tasaciones
-          </p>
-          <p className="mt-6 text-lg leading-8 text-gray-600">
-            Puedes dejar tu peticion para que hagamos una tasacion de tu
-            propiedad, a la brevedad nos contactaremos contigo para ultimar
-            detalles.
-          </p>
-        </div>
+    <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
+      Formulario Tasaciones
+    </p>
+    <p className="mt-6 text-lg leading-8 text-gray-600">
+      Puedes dejar tu peticion para que hagamos una tasacion de tu
+      propiedad, a la brevedad nos contactaremos contigo para ultimar
+      detalles.
+    </p>
+  </div>
+
+  )
+   
 }
