+++ client/src/app/Components/CardsConteiner/CardsConteiner.module.css
@@ -92,4 +92,3 @@ background-color: rgba(255, 255, 255, 0.20);
 .il {
   list-style-type: none;
 }
-
