+++ client/src/app/Contact/title.jsx
@@ -1,15 +1,14 @@
 
 export default function Title(){
-    <div className="mx-auto max-w-2xl lg:text-center">
-    <h2 className="text-base font-semibold leading-7 text-indigo-600">
-      Nosotros
-    </h2>
-    <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-      Formulario Contacto
-    </p>
-    <p className="mt-6 text-lg leading-8 text-gray-600">
-      Puedes dejar tu consulta, junto a los datos requeridos, a la
-      brevedad nos comunicaremos contigo.
-    </p>
-  </div>
+  return(<div className="mx-auto max-w-2xl lg:text-center">
+  
+  <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
+    Formulario Contacto
+  </p>
+  <p className="mt-6 text-lg leading-8 text-gray-600">
+    Puedes dejar tu consulta, junto a los datos requeridos, a la
+    brevedad nos comunicaremos contigo.
+  </p>
+</div>)
+    
 }
